feat(IAM-293): setting up external data source for page
